+++ tests/testthat/test_plots.R
@@ -11,6 +11,93 @@ yeats <- data.frame(
   angle = c(0, 315, 270, 225, 180, 135, 90, 45, 0, 315, 270)
 )
 
+context("geom_bar_text()")
+
+test_that("plots look the way they should", {
+
+  vdiffr::expect_doppelganger("coord_flip", {
+    ggplot(altitudes, aes(x = craft, y = altitude, label = altitude)) +
+      coord_flip() +
+      geom_col() +
+      geom_bar_text()
+  })
+
+  vdiffr::expect_doppelganger("implied flip", {
+    ggplot(altitudes, aes(y = craft, x = altitude, label = altitude)) +
+      geom_col() +
+      geom_bar_text()
+  })
+
+  vdiffr::expect_doppelganger("coord_flip with grow", {
+    ggplot(altitudes, aes(x = craft, y = altitude, label = altitude)) +
+      coord_flip() +
+      geom_col() +
+      geom_bar_text(grow = TRUE)
+  })
+
+  vdiffr::expect_doppelganger("implied flip with grow", {
+    ggplot(altitudes, aes(y = craft, x = altitude, label = altitude)) +
+      geom_col() +
+      geom_bar_text(grow = TRUE)
+  })
+
+  vdiffr::expect_doppelganger("coord_flip with place middle", {
+    ggplot(altitudes, aes(x = craft, y = altitude, label = altitude)) +
+      coord_flip() +
+      geom_col() +
+      geom_bar_text(place = "middle")
+  })
+
+  vdiffr::expect_doppelganger("implied flip with place middle", {
+    ggplot(altitudes, aes(y = craft, x = altitude, label = altitude)) +
+      geom_col() +
+      geom_bar_text(place = "middle")
+  })
+
+  vdiffr::expect_doppelganger("coord_flip with dodge", {
+    ggplot(coffees, aes(x = coffee, y = proportion, label = ingredient,
+                        fill = ingredient)) +
+      geom_col(position = "dodge") +
+      geom_bar_text(position = "dodge") +
+      coord_flip()
+  })
+
+  vdiffr::expect_doppelganger("coord_flip with dodge and grow and reflow", {
+    ggplot(coffees, aes(x = coffee, y = proportion, label = ingredient,
+                        fill = ingredient)) +
+      geom_col(position = "dodge") +
+      geom_bar_text(position = "dodge", grow = TRUE, reflow = TRUE, 
+                    place = "left") +
+      coord_flip()
+  })
+
+  vdiffr::expect_doppelganger("implied flip with dodge", {
+    ggplot(coffees, aes(y = coffee, x = proportion, label = ingredient,
+                        fill = ingredient)) +
+      geom_col(position = "dodge") +
+      geom_bar_text(position = "dodge")
+  })
+
+  vdiffr::expect_doppelganger("implied flip with dodge and grow and reflow", {
+    ggplot(coffees, aes(y = coffee, x = proportion, label = ingredient,
+                        fill = ingredient)) +
+      geom_col(position = "dodge") +
+      geom_bar_text(position = "dodge", grow = TRUE, reflow = TRUE, 
+                    place = "left")
+  })
+
+
+  vdiffr::expect_doppelganger("implied doesn't misfire with integer y", {
+    c2 <- coffees
+    c2$proportion <- 1:6
+    ggplot(c2, aes(x = coffee, y = proportion, label = ingredient,
+                        fill = ingredient)) +
+    geom_col(position = "dodge") +
+    geom_bar_text(position = "dodge")
+  })
+
+})
+
 context("visual tests of plots")
 
 test_that("plots look the way they should", {